requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
airtable-python-wrapper==0.15.1
    # via -r requirements.in
brotli==1.0.9
    # via flask-compress
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via flask
dash-core-components==1.15.0
    # via dash
dash-dangerously-set-inner-html==0.0.2
    # via -r requirements.in
dash-html-components==1.1.2
    # via dash
dash-table==4.11.2
    # via dash
dash==1.19.0
    # via -r requirements.in
dash_renderer==1.9.0
    # via dash
flask-compress==1.9.0
    # via dash
flask==1.1.2
    # via
    #   dash
    #   flask-compress
future==0.18.2
    # via dash
idna==2.10
    # via requests
itsdangerous==1.1.0
    # via flask
jinja2==2.11.3
    # via flask
markupsafe==1.1.1
    # via jinja2
numpy==1.20.1
    # via pandas
pandas==1.2.3
    # via -r requirements.in
plotly==4.14.3
    # via dash
python-dateutil==2.8.1
    # via pandas
python-dotenv==0.15.0
    # via -r requirements.in
python-slugify==4.0.1
    # via -r requirements.in
pytz==2021.1
    # via pandas
requests==2.25.1
    # via airtable-python-wrapper
retrying==1.3.3
    # via plotly
six==1.15.0
    # via
    #   plotly
    #   python-dateutil
    #   retrying
text-unidecode==1.3
    # via python-slugify
urllib3==1.26.3
    # via requests
werkzeug==1.0.1
    # via flask
